// src/app/types/ShoppingActionType.enum.ts
export enum ShoppingActionType {
    LOAD_SHOPPING = '[SHOPPING] Load Shopping',
    LOAD_SHOPPING_SUCCESS = '[SHOPPING] Load Shopping Success',
    LOAD_SHOPPING_FAILURE = '[SHOPPING] Load Shopping Failure',
    ADD_ITEM = '[SHOPPING] Add Item',
    ADD_ITEM_SUCCESS = '[SHOPPING] Add Item Success',
    ADD_ITEM_FAILURE = '[SHOPPING] Add Item Failure',
    DELETE_ITEM = '[SHOPPING] Delete Item',
    DELETE_ITEM_SUCCESS = '[SHOPPING] Delete Item Success',
    DELETE_ITEM_FAILURE = '[SHOPPING] Delete Item Failure'
}